fix(appeals): use correct keys for developmentType - Move APPEAL_DEVELOPMENT_TYPES constant to shared package to keep it   consistent between web & API - Update APPEAL_DEVELOPMENT_TYPES keys to match the data coming from FO
A2-3902
