--- appeals/api/src/server/endpoints/appellant-cases/appellant-cases.constants.js
@@ -1,17 +1,0 @@
-export const APPEAL_DEVELOPMENT_TYPES = [
-	{ value: 'householder', label: 'householder developments' },
-	{ value: 'change-of-use', label: 'change of use' },
-	{ value: 'major-dwellings', label: 'major dwellings' },
-	{ value: 'major-general-industry', label: 'major general industry/storage/warehousing' },
-	{ value: 'major-offices', label: 'major offices/R&D/light industry' },
-	{ value: 'major-retail', label: 'major retail and services' },
-	{ value: 'major-traveller', label: 'major traveller and caravan pitches' },
-	{ value: 'mineral-working', label: 'mineral working' },
-	{ value: 'minor-dwellings', label: 'minor dwellings' },
-	{ value: 'minor-general-industry', label: 'minor general industry/storage/warehousing' },
-	{ value: 'minor-offices', label: 'minor offices/R&D/light industry' },
-	{ value: 'minor-retail', label: 'minor retail and services' },
-	{ value: 'minor-traveller', label: 'minor traveller and caravan pitches' },
-	{ value: 'other-major', label: 'other major development' },
-	{ value: 'other-minor', label: 'other minor developments' }
-];
