--- appeals/web/src/server/appeals/appeal-details/appellant-case/application-development-type/appeal-development-type.constants.js
@@ -1,17 +1,0 @@
-export const APPEAL_DEVELOPMENT_TYPES = [
-	{ value: 'householder', label: 'Householder developments' },
-	{ value: 'change-of-use', label: 'Change of use' },
-	{ value: 'major-dwellings', label: 'Major dwellings' },
-	{ value: 'major-general-industry', label: 'Major general industry/storage/warehousing' },
-	{ value: 'major-offices', label: 'Major offices/R&D/light industry' },
-	{ value: 'major-retail', label: 'Major retail and services' },
-	{ value: 'major-traveller', label: 'Major traveller and caravan pitches' },
-	{ value: 'mineral-working', label: 'Mineral working' },
-	{ value: 'minor-dwellings', label: 'Minor dwellings' },
-	{ value: 'minor-general-industry', label: 'Minor general industry/storage/warehousing' },
-	{ value: 'minor-offices', label: 'Minor offices/R&D/light industry' },
-	{ value: 'minor-retail', label: 'Minor retail and services' },
-	{ value: 'minor-traveller', label: 'Minor traveller and caravan pitches' },
-	{ value: 'other-major', label: 'Other major development' },
-	{ value: 'other-minor', label: 'Other minor developments' }
-];
