+++ appeals/web/src/server/appeals/appeal-details/appellant-case/application-development-type/application-development-type.mapper.js
@@ -2,7 +2,8 @@
  * @typedef {import('../../appeal-details.types.js').WebAppeal} Appeal
  */
 import { appealShortReference } from '#lib/appeals-formatter.js';
-import { APPEAL_DEVELOPMENT_TYPES } from './appeal-development-type.constants.js';
+import { capitalizeFirstLetter } from '#lib/string-utilities.js';
+import { APPEAL_DEVELOPMENT_TYPES } from '@pins/appeals/constants/appellant-cases.constants.js';
 
 /**
  * @param {Appeal} appealData
@@ -31,11 +32,13 @@ export const changeDevelopmentTypePage = (appealData, appellantCaseData, storedS
 							classes: 'govuk-fieldset__legend--l'
 						}
 					},
-					items: APPEAL_DEVELOPMENT_TYPES.map(({ value, label }) => ({
-						value,
-						text: label,
-						checked: value === selectedValue
-					}))
+					items: APPEAL_DEVELOPMENT_TYPES.map(
+						(/** @type {{value: string, label: string}} */ { value, label }) => ({
+							value,
+							text: capitalizeFirstLetter(label),
+							checked: value === selectedValue
+						})
+					)
 				}
 			}
 		]
