+++ appeals/web/src/server/appeals/appeal-details/appellant-case/application-development-type/application-development-type.validators.js
@@ -1,8 +1,10 @@
 import { body } from 'express-validator';
 import { createValidator } from '@pins/express';
-import { APPEAL_DEVELOPMENT_TYPES } from './appeal-development-type.constants.js';
+import { APPEAL_DEVELOPMENT_TYPES } from '@pins/appeals/constants/appellant-cases.constants.js';
 
-const allowedDevelopmentTypes = APPEAL_DEVELOPMENT_TYPES.map(({ value }) => value);
+const allowedDevelopmentTypes = APPEAL_DEVELOPMENT_TYPES.map(
+	(/** @type {{value: string}} */ { value }) => value
+);
 
 export const validateDevelopmentTypePresent = createValidator([
 	body('developmentType').exists({ checkFalsy: true }).withMessage('Select a development type'),
