+++ appeals/web/src/server/lib/mappers/data/appellant-case/submappers/development-type.js
@@ -1,11 +1,14 @@
-import { APPEAL_DEVELOPMENT_TYPES } from '#appeals/appeal-details/appellant-case/application-development-type/appeal-development-type.constants.js';
+import { APPEAL_DEVELOPMENT_TYPES } from '@pins/appeals/constants/appellant-cases.constants.js';
 import { textSummaryListItem } from '#lib/mappers/components/index.js';
+import { capitalizeFirstLetter } from '#lib/string-utilities.js';
 
 /** @type {import('../mapper.js').SubMapper} */
 export const mapDevelopmentType = ({ appellantCaseData, currentRoute, userHasUpdateCase }) => {
 	const code = appellantCaseData?.developmentType;
-	const entry = APPEAL_DEVELOPMENT_TYPES.find((item) => item.value === code);
-	const label = entry?.label || 'Not provided';
+	const entry = APPEAL_DEVELOPMENT_TYPES.find(
+		(/** @type {{value: string, label: string}} */ item) => item.value === code
+	);
+	const label = capitalizeFirstLetter(entry?.label || 'Not provided');
 
 	return textSummaryListItem({
 		id: 'development-type',
