+++ packages/appeals/constants/appellant-cases.constants.js
@@ -1,0 +1,18 @@
+/** @type {Array<{value: string, label: string}>} */
+export const APPEAL_DEVELOPMENT_TYPES = [
+	{ value: 'householder', label: 'householder developments' },
+	{ value: 'change-of-use', label: 'change of use' },
+	{ value: 'major-dwellings', label: 'major dwellings' },
+	{ value: 'major-industry-storage', label: 'major general industry/storage/warehousing' },
+	{ value: 'major-offices', label: 'major offices/R&D/light industry' },
+	{ value: 'major-retail-services', label: 'major retail and services' },
+	{ value: 'major-traveller-caravan', label: 'major traveller and caravan pitches' },
+	{ value: 'mineral-workings', label: 'mineral working' },
+	{ value: 'minor-dwellings', label: 'minor dwellings' },
+	{ value: 'minor-industry-storage', label: 'minor general industry/storage/warehousing' },
+	{ value: 'minor-offices', label: 'minor offices/R&D/light industry' },
+	{ value: 'minor-retail-services', label: 'minor retail and services' },
+	{ value: 'minor-traveller-caravan', label: 'minor traveller and caravan pitches' },
+	{ value: 'other-major', label: 'other major development' },
+	{ value: 'other-minor', label: 'other minor developments' }
+];
