+++ packages/appeals/constants/index.d.ts
@@ -1,3 +1,4 @@
+declare module '@pins/appeals/constants/appellant-cases.constants.js';
 declare module '@pins/appeals/constants/common.js';
 declare module '@pins/appeals/constants/documents.js';
 declare module '@pins/appeals/constants/dates.js';
